link to portal documentation
The Nextcloud App Store [listing](https://apps.nextcloud.com/apps/user_saml) for this app continues to point to the URL of the last-seen `<documentation>` tag, which is a dead link.  That behavior is probably a bug in App Store, but we can overwrite it with an updated URL to reduce confusion.