Implement proper group mapping via SAML
Fixes #561

As stated in the issue, it's not desirable to have a group called
`admin` in the SAML backend which doesn't indicate to which service
admin permissions are granted.

This is orthogonal to `saml-attribute-mapping-group_mapping` which
simply maps all groups from a SAML attribute to Nextcloud groups, i.e.
the attribute's value MUST contain a group called `admin` to make sure
that users get admin rights in Nextcloud.

When enabled, the name of (another) attribute must be specified which
contains a list of SAML-specific groups, e.g.

    ["nextcloud-admins", "nextcloud-marketing"]

that can be mapped to e.g.

    ["admin", "marketing"]

Signed-off-by: Maximilian Bosch <maximilian@mbosch.me>