fix(Session): restore session data after cookie login
when a browser is closed and the session lost, a new session might be
started with the help of the remember me cookie. For we keep some data in
the session, like the IdP identifier, and some SAML data needed for SLO,
those were lost in this process. This made especially logout behaviour
not acting as expected.

Now we keep the required data also in the database and can restore it on
a cookie login event. Storing the session data in the database can only
happen after the session token was created, i.e. when the UserLoggedInEvent
was dispatched.

Stored session data is deleted once the original authtoken has also
disappeared. A daily check during maintenance times is scheduled.

Because hashing of the session id is a private implementation detail of
the current token provider, we track the id of the token as well so that
any future change in hashing will not have any effect.

Signed-off-by: Arthur Schiwon <blizzz@arthur-schiwon.de>